# news-aggregatorcomposer/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API Documentation",
        "description": "News Aggregator API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:70/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Fetch articles with pagination",
                "operationId": "a18fe99eedbd33cdb4dbbb8b5dac8bb4",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Fetch a specific article by ID",
                "operationId": "126f0079c2def3834cf2c4a2456d16f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "A child needed antivenom for a snakebite. It cost more than $200,000. - The Washington Post"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "It took more than 30 vials of antivenom to save a toddler’s life after he was bitten by a rattlesnake."
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "example": "https://www.washingtonpost.com/wellness/2024/10/30/snakebite-antivenom-child-cost-bill/"
                                                },
                                                "url_to_image": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "example": "https://www.washingtonpost.com/wp-apps/imrs.php?src=https://arc-anglerfish-washpost-prod-washpost.s3.amazonaws.com/public/GJPDFSKQ5GTXARYX5B5JAL4VBA_size-normalized.JPG&w=1440"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Jackie Fortiér"
                                                },
                                                "source_name": {
                                                    "type": "string",
                                                    "example": "The Washington Post"
                                                },
                                                "source_id": {
                                                    "type": "string",
                                                    "example": "the-washington-post"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-30 12:40:30"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "A few days after his 2nd birthday, Brigland Pfeffer was playing with his siblings..."
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T11:43:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T11:43:49.000000Z"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}